requirements.txt
numpy
pandas
matplotlib.pyplot
seaborn
nltk
nltk.stem
sklearn
sklearn.feature_extraction.text
sklearn.metrics.pairwise